Move new Tokenizer classes into a TextMate namespace
