+++ src/TextMate/LineTokens.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phiki;
+namespace Phiki\TextMate;
 
 use Phiki\Token\Token;
