+++ src/TextMate/StateStack.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phiki;
+namespace Phiki\TextMate;
 
 use Phiki\Contracts\PatternInterface;
 use Phiki\Grammar\EndPattern;
